requirements.txt
requests~=2.32.5
bs4~=0.0.2
beautifulsoup4~=4.13.5
readability-lxml~=0.8.4.1
pandas~=2.3.2
google-search-results
